+++ example/lib/main.dart
@@ -56,8 +56,7 @@ class _HomePageState extends State<HomePage> {
         EasyTableColumn(
             name: 'Skills',
             width: 100,
-            cellBuilder: (context, row) =>
-                EasyTableCell(child: SkillsWidget(skills: row.skills))),
+            cellBuilder: (context, row) => SkillsWidget(skills: row.skills)),
         EasyTableColumn(
             name: 'Strength', width: 80, intValue: (row) => row.strength),
         EasyTableColumn(